virttest.utils_test.qemu: move migration codes to a single file 1. create folder "qemu" under utils_test 2. move qemu.py to folder "qemu" 3. move migration codes from qemu.py to migration.py 4. rename qemu.py to __init__.py
