+++ virttest/utils_test/qemu/migration.py
@@ -1,19 +1,7 @@
 """
-High-level QEMU test utility functions.
+High-level migration test utility functions.
 
-This module is meant to reduce code size by performing common test procedures.
-Generally, code here should look like test code.
-
-More specifically:
-    - Functions in this module should raise exceptions if things go wrong
-    - Functions in this module typically use functions and classes from
-      lower-level modules (e.g. utils_misc, qemu_vm, aexpect).
-    - Functions in this module should not be used by lower-level modules.
-    - Functions in this module should be used in the right context.
-      For example, a function should not be used where it may display
-      misleading or inaccurate info or debug messages.
-
-:copyright: 2008-2013 Red Hat Inc.
+This module includes framework and some public functions.
 """
 
 import cPickle
@@ -21,27 +9,27 @@ import errno
 import fcntl
 import logging
 import os
-import re
 import socket
 import threading
 import time
-from functools import reduce
-
+import re
 from avocado.core import exceptions
 from avocado.utils import crypto
 from avocado.utils import data_factory
 from avocado.utils import path as utils_path
 from avocado.utils import process
-
-from .. import data_dir
-from .. import env_process
-from .. import error_context
-from .. import remote
-from .. import storage
-from .. import utils_misc
-from .. import qemu_monitor
-from ..qemu_devices import qdevices
-from ..staging import utils_memory
+from virttest import data_dir
+from virttest import remote
+from virttest import storage
+from virttest import utils_test
+from virttest import utils_misc
+from virttest import env_process
+from autotest.client.shared import error
+from autotest.client.shared import utils
+try:
+    import aexpect
+except ImportError:
+    from virttest import aexpect
 
 
 def guest_active(vm):
@@ -55,40 +43,9 @@ def guest_active(vm):
             return o.get("running")
 
 
-def get_numa_status(numa_node_info, qemu_pid, debug=True):
-    """
-    Get the qemu process memory use status and the cpu list in each node.
-
-    :param numa_node_info: Host numa node information
-    :type numa_node_info: NumaInfo object
-    :param qemu_pid: process id of qemu
-    :type numa_node_info: string
-    :param debug: Print the debug info or not
-    :type debug: bool
-    :return: memory and cpu list in each node
-    :rtype: tuple
-    """
-    node_list = numa_node_info.online_nodes
-    qemu_memory = []
-    qemu_cpu = []
-    cpus = utils_misc.get_pid_cpu(qemu_pid)
-    for node_id in node_list:
-        qemu_memory_status = utils_memory.read_from_numa_maps(qemu_pid,
-                                                              "N%d" % node_id)
-        memory = sum([int(_) for _ in qemu_memory_status.values()])
-        qemu_memory.append(memory)
-        cpu = [_ for _ in cpus if _ in numa_node_info.nodes[node_id].cpus]
-        qemu_cpu.append(cpu)
-        if debug:
-            logging.debug("qemu-kvm process using %s pages and cpu %s in "
-                          "node %s" % (memory, " ".join(cpu), node_id))
-    return (qemu_memory, qemu_cpu)
-
-
 def get_nic_vendor(params, cmd):
     """
     Get host link layer
-
     :param params: Dictionary with the test parameters.
     :param cmd: Command string
     """
@@ -105,129 +62,6 @@ def get_nic_vendor(params, cmd):
     if nic_vendor not in expected_nic_vendor.split():
         raise exceptions.TestError("The Link layer is not correct, "
                                    "expected is '%s'" % expected_nic_vendor)
-
-
-def pin_vm_threads(vm, node):
-    """
-    Pin VM threads to single cpu of a numa node
-
-    :param vm: VM object
-    :param node: NumaNode object
-    """
-    if len(vm.vcpu_threads) + len(vm.vhost_threads) < len(node.cpus):
-        for i in vm.vcpu_threads:
-            logging.info("pin vcpu thread(%s) to cpu(%s)" %
-                         (i, node.pin_cpu(i)))
-        for i in vm.vhost_threads:
-            logging.info("pin vhost thread(%s) to cpu(%s)" %
-                         (i, node.pin_cpu(i)))
-    elif (len(vm.vcpu_threads) <= len(node.cpus) and
-          len(vm.vhost_threads) <= len(node.cpus)):
-        for i in vm.vcpu_threads:
-            logging.info("pin vcpu thread(%s) to cpu(%s)" %
-                         (i, node.pin_cpu(i)))
-        for i in vm.vhost_threads:
-            logging.info("pin vhost thread(%s) to extra cpu(%s)" %
-                         (i, node.pin_cpu(i, extra=True)))
-    else:
-        logging.info("Skip pinning, no enough nodes")
-
-
-def _check_driver_verifier(session, driver, timeout=300):
-    """
-    Check driver verifier status
-
-    :param session: VM session.
-    :param driver: The driver need to query
-    :param timeout: Timeout in seconds
-    """
-    logging.info("Check %s driver verifier status" % driver)
-    query_cmd = "verifier /querysettings"
-    output = session.cmd_output(query_cmd, timeout=timeout)
-    return (driver in output, output)
-
-
-@error_context.context_aware
-def setup_win_driver_verifier(driver, vm, timeout=300):
-    """
-    Enable driver verifier for windows guest.
-
-    :param driver: The driver which needs enable the verifier.
-    :param vm: VM object.
-    :param timeout: Timeout in seconds.
-    """
-    session = vm.wait_for_login(timeout=timeout)
-    try:
-        verifier_status = _check_driver_verifier(session, driver)[1]
-        if not verifier_status:
-            error_context.context("Enable %s driver verifier" % driver,
-                                  logging.info)
-            verifier_setup_cmd = "verifier /standard /driver %s.sys" % driver
-            session.cmd(verifier_setup_cmd,
-                        timeout=timeout,
-                        ignore_all_errors=True)
-            session = vm.reboot(session)
-            verifier_status, output = _check_driver_verifier(session, driver)
-            if not verifier_status:
-                msg = "%s verifier is not enabled, details: %s" % (driver,
-                                                                   output)
-                raise exceptions.TestFail(msg)
-        logging.info("%s verifier is enabled already" % driver)
-    finally:
-        session.close()
-
-
-def clear_win_driver_verifier(driver, vm, timeout=300):
-    """
-    Clear the driver verifier in windows guest.
-
-    :param driver: The driver need to clear
-    :param vm: VM object.
-    :param timeout: Timeout in seconds.
-    """
-    session = vm.wait_for_login(timeout=timeout)
-    try:
-        verifier_status = _check_driver_verifier(session, driver)[1]
-        if verifier_status:
-            logging.info("Clear driver verifier")
-            verifier_clear_cmd = "verifier /reset"
-            session.cmd(verifier_clear_cmd,
-                        timeout=timeout,
-                        ignore_all_errors=True)
-            session = vm.reboot(session)
-    finally:
-        session.close()
-
-
-def setup_runlevel(params, session):
-    """
-    Setup the runlevel in guest.
-
-    :param params: Dictionary with the test parameters.
-    :param session: VM session.
-    """
-    cmd = "runlevel"
-    ori_runlevel = "0"
-    expect_runlevel = params.get("expect_runlevel", "3")
-
-    # Note: All guest services may have not been started when
-    #       the guest gets IP addr; the guest runlevel maybe
-    #       is "unknown" whose exit status is 1 at that time,
-    #       which will cause the cmd execution failed. Need some
-    #       time here to wait for the guest services start.
-    if utils_misc.wait_for(lambda: session.cmd_status(cmd) == 0, 15):
-        ori_runlevel = session.cmd(cmd)
-
-    ori_runlevel = ori_runlevel.split()[-1]
-    if ori_runlevel == expect_runlevel:
-        logging.info("Guest runlevel is already %s as expected" % ori_runlevel)
-    else:
-        session.cmd("init %s" % expect_runlevel)
-        tmp_runlevel = session.cmd(cmd)
-        tmp_runlevel = tmp_runlevel.split()[-1]
-        if tmp_runlevel != expect_runlevel:
-            logging.warn("Changing runlevel from %s to %s failed (%s)!" %
-                         ori_runlevel, expect_runlevel, tmp_runlevel)
 
 
 def migrate(vm, env=None, mig_timeout=3600, mig_protocol="tcp",
@@ -1377,468 +1211,459 @@ class MultihostMigrationRdma(MultihostMigration):
         utils_misc.parallel(multi_mig)
 
 
-class GuestSuspend(object):
+class MigrationBase(object):
 
-    """
-    Suspend guest, supports both Linux and Windows.
+    """Class that provides some general functions for multi-host migration."""
 
-    """
-    SUSPEND_TYPE_MEM = "mem"
-    SUSPEND_TYPE_DISK = "disk"
+    def __setup__(self, test, params, env, srchost, dsthost):
 
-    def __init__(self, params, vm):
-        if not params or not vm:
-            raise exceptions.TestError("Missing 'params' or 'vm' parameters")
-
-        self._open_session_list = []
-        self.vm = vm
-        self.params = params
-        self.login_timeout = float(self.params.get("login_timeout", 360))
-        self.services_up_timeout = float(self.params.get("services_up_timeout",
-                                                         30))
-        self.os_type = self.params.get("os_type")
-
-    def _get_session(self):
-        self.vm.verify_alive()
-        session = self.vm.wait_for_login(timeout=self.login_timeout)
-        return session
-
-    def _session_cmd_close(self, session, cmd):
-        try:
-            return session.cmd_status_output(cmd)
-        finally:
-            try:
-                session.close()
-            except Exception:
-                pass
-
-    def _cleanup_open_session(self):
-        try:
-            for s in self._open_session_list:
-                if s:
-                    s.close()
-        except Exception:
-            pass
-
-    @error_context.context_aware
-    def setup_bg_program(self, **args):
+        """initialize some public params
         """
-        Start up a program as a flag in guest.
-        """
-        suspend_bg_program_setup_cmd = args.get("suspend_bg_program_setup_cmd")
 
-        error_context.context(
-            "Run a background program as a flag", logging.info)
-        session = self._get_session()
-        self._open_session_list.append(session)
-
-        logging.debug("Waiting all services in guest are fully started.")
-        time.sleep(self.services_up_timeout)
-
-        session.sendline(suspend_bg_program_setup_cmd)
-
-    @error_context.context_aware
-    def check_bg_program(self, **args):
-        """
-        Make sure the background program is running as expected
-        """
-        suspend_bg_program_chk_cmd = args.get("suspend_bg_program_chk_cmd")
-
-        error_context.context(
-            "Verify background program is running", logging.info)
-        session = self._get_session()
-        s, _ = self._session_cmd_close(session, suspend_bg_program_chk_cmd)
-        if s:
-            raise exceptions.TestFail(
-                "Background program is dead. Suspend failed.")
-
-    @error_context.context_aware
-    def kill_bg_program(self, **args):
-        error_context.context("Kill background program after resume")
-        suspend_bg_program_kill_cmd = args.get("suspend_bg_program_kill_cmd")
-
-        try:
-            session = self._get_session()
-            self._session_cmd_close(session, suspend_bg_program_kill_cmd)
-        except Exception, e:
-            logging.warn("Could not stop background program: '%s'", e)
-            pass
-
-    @error_context.context_aware
-    def _check_guest_suspend_log(self, **args):
-        error_context.context("Check whether guest supports suspend",
-                              logging.info)
-        suspend_support_chk_cmd = args.get("suspend_support_chk_cmd")
-
-        session = self._get_session()
-        s, o = self._session_cmd_close(session, suspend_support_chk_cmd)
-
-        return s, o
-
-    def verify_guest_support_suspend(self, **args):
-        s, _ = self._check_guest_suspend_log(**args)
-        if s:
-            raise exceptions.TestError("Guest doesn't support suspend.")
-
-    @error_context.context_aware
-    def start_suspend(self, **args):
-        suspend_start_cmd = args.get("suspend_start_cmd")
-        error_context.context(
-            "Start suspend [%s]" % (suspend_start_cmd), logging.info)
-
-        session = self._get_session()
-        self._open_session_list.append(session)
-
-        # Suspend to disk
-        session.sendline(suspend_start_cmd)
-
-    @error_context.context_aware
-    def verify_guest_down(self, **args):
-        # Make sure the VM goes down
-        error_context.context("Wait for guest goes down after suspend")
-        suspend_timeout = 240 + int(self.params.get("smp")) * 60
-        if not utils_misc.wait_for(self.vm.is_dead, suspend_timeout, 2, 2):
-            raise exceptions.TestFail("VM refuses to go down. Suspend failed.")
-
-    @error_context.context_aware
-    def resume_guest_mem(self, **args):
-        error_context.context("Resume suspended VM from memory")
-        self.vm.monitor.system_wakeup()
-
-    @error_context.context_aware
-    def resume_guest_disk(self, **args):
-        error_context.context("Resume suspended VM from disk")
-        self.vm.create()
-
-    @error_context.context_aware
-    def verify_guest_up(self, **args):
-        error_context.context("Verify guest system log", logging.info)
-        suspend_log_chk_cmd = args.get("suspend_log_chk_cmd")
-
-        session = self._get_session()
-        s, o = self._session_cmd_close(session, suspend_log_chk_cmd)
-        if s:
-            raise exceptions.TestError(
-                "Could not find suspend log. [%s]" % (o))
-
-    @error_context.context_aware
-    def action_before_suspend(self, **args):
-        error_context.context("Actions before suspend")
-        pass
-
-    @error_context.context_aware
-    def action_during_suspend(self, **args):
-        error_context.context(
-            "Sleep a while before resuming guest", logging.info)
-
-        time.sleep(10)
-        if self.os_type == "windows":
-            # Due to WinXP/2003 won't suspend immediately after issue S3 cmd,
-            # delay 10~60 secs here, maybe there's a bug in windows os.
-            logging.info("WinXP/2003 need more time to suspend, sleep 50s.")
-            time.sleep(50)
-
-    @error_context.context_aware
-    def action_after_suspend(self, **args):
-        error_context.context("Actions after suspend")
-        pass
-
-
-class MemoryBaseTest(object):
-
-    """
-    Base class for memory functions.
-    """
-
-    UNIT = "M"
-
-    def __init__(self, test, params, env):
         self.test = test
-        self.env = env
         self.params = params
-        self.sessions = {}
+        self.env = env
+        self.srchost = srchost
+        self.dsthost = dsthost
+        self.vms = params.objects("vms")
+        self.vm = self.vms[0]
+        self.is_src = params["hostid"] == self.srchost
+        self.pre_sub_test = params.get("pre_sub_test")
+        self.post_sub_test = params.get("post_sub_test")
+        self.login_before_pre_tests = params.get("login_before_pre_tests",
+                                                 "no")
+        self.mig_bg_command = params.get("migration_bg_command",
+                                         "cd /tmp; nohup ping localhost &")
+        self.mig_bg_check_command = params.get("migration_bg_check_command",
+                                               "pgrep ping")
+        self.mig_bg_kill_command = params.get("migration_bg_kill_command",
+                                              "pkill -9 ping")
+        self.migration_timeout = int(params.get("migration_timeout",
+                                                "1500"))
+        self.login_timeout = 480
+        self.stop_migrate = False
+        self.migrate_count = int(params.get("migrate_count", 1))
+        self.id = {"src": self.srchost,
+                   "dst": self.dsthost,
+                   "type": "file_transfer"}
+        self.capabilitys = params.objects("capabilitys")
+        self.capabilitys_state = params.objects("capabilitys_state")
+        for i in range(0, len(self.capabilitys_state)):
+            if self.capabilitys_state[i].strip() == "enable":
+                self.capabilitys_state[i] = True
+            else:
+                self.capabilitys_state[i] = False
+        self.parameters = params.objects("parameters")
+        self.parameters_value = params.objects("parameters_value")
+        self.cache_size = params.objects("cache_size")
+        self.kill_bg_stress_cmd = params.get("kill_bg_stress_cmd",
+                                             "killall -9 stress")
+        self.bg_stress_test = params.get("bg_stress_test")
+        self.check_running_cmd = params.get("check_running_cmd")
+        self.max_speed = params.get("max_migration_speed", "1000")
+        self.max_speed = utils.convert_data_size(self.max_speed, "M")
+        self.need_set_speed = params.get("need_set_speed", "yes") == "yes"
+        self.WAIT_SHORT = 15
 
-    def get_vm_mem(self, vm):
+    @error.context_aware
+    def run_pre_sub_test(self):
+
         """
-        Count memory assigned to VM.
-
-        :param vm: VM object
-        :return: memory size in MB.
+        run sub test on src before migration
         """
-        PC_DIMM = qdevices.Dimm
-        # Default memory size in configuration is MB, so append
-        # 'MB' in the end of 'mem' param.
-        mem_str = "%sM" % vm.params.get("mem", "0")
-        total_mem = self.normalize_mem_size(mem_str)
-        pc_dimms = filter(lambda x: isinstance(x, PC_DIMM), vm.devices)
-        obj_ids = map(lambda x: x.get_param('memdev'), pc_dimms)
-        obj_devs = map(lambda x: vm.devices.get_by_qid(x)[0], obj_ids)
-        obj_size = map(lambda x: x.get_param('size'), obj_devs)
-        total_mem += sum(map(self.normalize_mem_size, obj_size))
-        logging.info("Assigned %s%s " % (total_mem, self.UNIT) +
-                     "memory to '%s'" % vm.name)
-        return total_mem
 
-    @classmethod
-    def normalize_mem_size(cls, str_size):
+        if self.is_src:
+            if self.pre_sub_test:
+                if self.login_before_pre_tests == "yes":
+                    vm = self.env.get_vm(self.params["main_vm"])
+                    vm.wait_for_login(timeout=self.login_timeout)
+                error.context("Run sub test '%s' before migration on src"
+                              % self.pre_sub_test, logging.info)
+                utils_test.run_virt_sub_test(self.test, self.params,
+                                             self.env, self.pre_sub_test)
+
+    @error.context_aware
+    def run_post_sub_test(self):
+
         """
-        Convert memory size unit
-
-        :param str_size: memory size string, like: 1GB
-        :return: memory size value in MB
+        run sub test on dst after migration
         """
-        args = (str_size, cls.UNIT, 1024)
-        try:
-            size = utils_misc.normalize_data_size(*args)
-            return int(float(size))
-        except ValueError, details:
-            logging.debug("Convert memory size error('%s')" % details)
-        return 0
 
-    @classmethod
-    def get_guest_total_mem(cls, vm):
+        if not self.is_src:
+            if self.post_sub_test:
+                error.context("Run sub test '%s' after migration on dst"
+                              % self.post_sub_test, logging.info)
+                utils_test.run_virt_sub_test(self.test, self.params,
+                                             self.env, self.post_sub_test)
+
+    def prepare_vm(self, vm_name):
+
         """
-        Guest OS reported physical memory size in MB.
-
-        :param vm: VM object.
-        :return: physical memory report by guest OS in MB
+        Prepare, start vm and return vm.
+        :param vm_name: vm name to be started.
+        :return: Started VM.
         """
-        if vm.params.get("os_type") == "windows":
-            cmd = 'wmic ComputerSystem get TotalPhysicalMemory'
-        else:
-            cmd = "grep 'MemTotal:' /proc/meminfo"
-        return vm.get_memory_size(cmd)
 
-    @classmethod
-    def get_guest_free_mem(cls, vm):
+        self.vm_lock = threading.Lock()
+        new_params = self.params.copy()
+        new_params['migration_mode'] = None
+        new_params['start_vm'] = 'yes'
+        self.vm_lock.acquire()
+        env_process.process(self.test, new_params, self.env,
+                            env_process.preprocess_image,
+                            env_process.preprocess_vm)
+        self.vm_lock.release()
+        vm = self.env.get_vm(vm_name)
+        vm.wait_for_login(timeout=self.login_timeout)
+        return vm
+
+    def start_worker(self):
+
         """
-        Guest OS reported free memory size in MB.
-
-        :param vm: VM Object
-        :return: free memory report by guest OS in MB
+        run background command on src before migration
         """
-        os_type = vm.params.get("os_type")
-        timeout = float(vm.params.get("login_timeout", 600))
-        session = vm.wait_for_login(timeout=timeout)
-        return utils_misc.get_free_mem(session, os_type)
 
-    def get_session(self, vm):
-        """
-        Get connection to VM.
-
-        :param vm: VM object
-        :return: return ShellSession object
-        """
-        key = vm.instance
-        self.sessions.setdefault(key, [])
-        for session in self.sessions.get(key):
-            if session.is_responsive():
-                return session
-            session.close()
-            self.sessions[key].remove(session)
-        login_timeout = float(self.params.get("login_timeout", 600))
-        session = vm.wait_for_login(timeout=login_timeout)
-        self.sessions[key].append(session)
-        return session
-
-    def close_sessions(self):
-        """
-        Close opening session, better to call it in the end of test.
-        """
-        sessions = filter(None, self.sessions.values())
-        if sessions:
-            sessions = filter(None, reduce(list.__add__, sessions))
-            map(lambda x: x.close(), sessions)
-        self.sessions.clear()
-
-
-class MemoryHotplugTest(MemoryBaseTest):
-
-    """
-    Class for memory hotplug/unplug test.
-    """
-
-    @error_context.context_aware
-    def update_vm_after_hotplug(self, vm, dev):
-        """
-        Update VM params to ensure hotpluged devices exist in guest.
-
-        :param vm: VM object
-        :param dev: Qdevice object.
-        """
-        error_context.context("Update VM object after hotplug memory")
-        attrs = dev.__attributes__[:]
-        params = self.params.copy_from_keys(attrs)
-        dev_type, name = dev.get_qid().split('-')
-        for attr in attrs:
-            val = dev.get_param(attr)
-            if val:
-                key = "_".join([attr, dev_type, name])
-                params[key] = val
-        mem_devs = vm.params.objects("mem_devs")
-        if name not in mem_devs:
-            mem_devs.append(name)
-            params["mem_devs"] = " ".join(mem_devs)
-        vm.params.update(params)
-        if dev not in vm.devices:
-            vm.devices.insert(dev)
-        self.env.register_vm(vm.name, vm)
-
-    @error_context.context_aware
-    def update_vm_after_unplug(self, vm, dev):
-        """
-        Update VM params object after unplug memory devices.
-
-        :param vm: VM object
-        :param dev: Qdevice object
-        """
-        error_context.context("Update VM object after unplug memory")
-        dev_type, name = dev.get_qid().split('-')
-        mem_devs = vm.params.objects("mem_devs")
-        if dev in vm.devices:
-            vm.devices.remove(dev)
-        if name in mem_devs:
-            mem_devs.remove(name)
-            vm.params["mem_devs"] = " ".join(mem_devs)
-        self.env.register_vm(vm.name, vm)
-
-    @error_context.context_aware
-    def hotplug_memory(self, vm, name):
-        """
-        Hotplug dimm device with memory backend
-
-        :param vm: VM object
-        :param name: memory device name
-        """
-        devices = vm.devices.memory_define_by_params(self.params, name)
-        for dev in devices:
-            dev_type = "memory"
-            if isinstance(dev, qdevices.Dimm):
-                addr = self.get_mem_addr(vm, dev.get_qid())
-                dev.set_param("addr", addr)
-                dev_type = "pc-dimm"
-            step = "Hotplug %s '%s' to VM" % (dev_type, dev.get_qid())
-            error_context.context(step, logging.info)
-            vm.devices.simple_hotplug(dev, vm.monitor)
-            self.update_vm_after_hotplug(vm, dev)
-        return devices
-
-    @error_context.context_aware
-    def unplug_memory(self, vm, name):
-        """
-        Unplug memory device
-        step 1, unplug memory object
-        step 2, unplug dimm device
-
-        :param vm: VM object
-        :param name: memory device name
-        """
-        devices = []
-        qid_mem = "mem-%s" % name
-        step = "Unplug memory object '%s'" % qid_mem
-        error_context.context(step, logging.info)
-        try:
-            mem = vm.devices.get_by_qid(qid_mem)[0]
-        except IndexError:
-            output = vm.monitor.query("memory-devices")
-            logging.debug("Memory devices: %s" % output)
-            msg = "Memory object '%s' not exists" % qid_mem
-            raise exceptions.TestError(msg)
-        try:
-            qid_dimm = "dimm-%s" % name
-            dimm = vm.devices.get_by_qid(qid_dimm)[0]
-        except IndexError:
-            logging.warn("'%s' is not used by any dimm" % qid_mem)
-        step = "Unplug pc-dimm '%s'" % qid_dimm
-        error_context.context(step, logging.info)
-        vm.devices.simple_unplug(dimm, vm.monitor)
-        devices.append(dimm)
-        self.update_vm_after_unplug(vm, dimm)
-        error_context.context(step, logging.info)
-        vm.devices.simple_unplug(mem, vm.monitor)
-        devices.append(mem)
-        self.update_vm_after_unplug(vm, mem)
-        return devices
-
-    @error_context.context_aware
-    def get_mem_addr(self, vm, qid):
-        """
-        Get guest memory address from qemu monitor
-
-        :param vm: VM object
-        :param qid: memory device qid
-        """
-        error_context.context("Get hotpluged memory address", logging.info)
-        if not isinstance(vm.monitor, qemu_monitor.QMPMonitor):
-            raise NotImplementedError
-        for info in vm.monitor.info("memory-devices"):
-            if str(info['data']['id']) == qid:
-                address = info['data']['addr']
-                logging.info("Memory address: %s" % address)
-                return address
-
-    @error_context.context_aware
-    def check_memory(self, vm=None):
-        """
-        Check is guest memory is really match assgined to VM.
-
-        :param vm: VM object, get VM object from env if vm is None.
-        """
-        error_context.context("Verify memory info", logging.info)
-        if not vm:
+        if self.is_src:
+            logging.info("Try to login guest before migration test.")
             vm = self.env.get_vm(self.params["main_vm"])
-        vm.verify_alive()
-        threshold = float(self.params.get("threshold", 0.10))
-        timeout = float(self.params.get("wait_resume_timeout", 60))
-        # Notes:
-        #    some sub test will pause VM, here need to wait VM resume
-        # then check memory info in guest.
-        utils_misc.wait_for(lambda: not vm.is_paused(), timeout=timeout)
-        utils_misc.verify_host_dmesg()
-        self.os_type = self.params.get("os_type")
-        guest_mem_size = super(MemoryHotplugTest, self).get_guest_total_mem(vm)
-        vm_mem_size = self.get_vm_mem(vm)
-        if abs(guest_mem_size - vm_mem_size) > vm_mem_size * threshold:
-            msg = ("Assigned '%s MB' memory to '%s'"
-                   "but, '%s MB' memory detect by OS" %
-                   (vm_mem_size, vm.name, guest_mem_size))
-            raise exceptions.TestFail(msg)
+            session = vm.wait_for_login(timeout=self.login_timeout)
+            logging.debug("Sending command: '%s'" % self.mig_bg_command)
+            s, o = session.cmd_status_output(self.mig_bg_command)
+            if s != 0:
+                raise error.TestError("Failed to run bg cmd in guest,"
+                                      " Output is '%s'." % o)
+            time.sleep(5)
 
-    @error_context.context_aware
-    def memory_operate(self, vm, memory, operation='online'):
-        error_context.context(
-            "%s %s in guest OS" %
-            (operation, memory), logging.info)
-        mem_sys_path = "/sys/devices/system/memory/%s" % memory
-        mem_state_path = os.path.join(mem_sys_path, 'state')
-        session = self.get_session(vm)
-        session.cmd("echo '%s' > %s" % (operation, mem_state_path))
-        output = session.cmd_output_safe("cat %s" % mem_state_path)
-        if operation not in output:
-            return exceptions.TestFail("Fail to %s %s" % (operation, memory))
+    def check_worker(self):
 
-    def get_memory_state(self, vm, memory):
-        """Get memorys state in guest OS"""
-        mem_sys_path = "/sys/devices/system/memory/%s" % memory
-        mem_state_path = os.path.join(mem_sys_path, 'state')
-        session = self.get_session(vm)
-        status, output = session.cmd_status_output("cat %s" % mem_state_path)
-        if status != 0:
-            raise exceptions.TestError("Fail to read %s state" % memory)
-        return output.strip()
+        """
+        check background command on dst after migration
+        """
 
-    def get_offline_memorys(self, vm):
-        """Get unusable memory in guest OS"""
-        def is_offline_memory(x):
-            return self.get_memory_state(vm, x) == 'offline'
+        if not self.is_src:
+            logging.info("Try to login guest after migration test.")
+            vm = self.env.get_vm(self.params["main_vm"])
+            serial_login = self.params.get("serial_login")
+            if serial_login == "yes":
+                session = vm.wait_for_serial_login(timeout=self.login_timeout)
+            else:
+                session = vm.wait_for_login(timeout=self.login_timeout)
+            logging.info("Check the background command in the guest.")
+            s, o = session.cmd_status_output(self.mig_bg_check_command)
+            if s:
+                raise error.TestFail("Background command not found,"
+                                     " Output is '%s'." % o)
+            logging.info("Kill the background command in the guest.")
+            session.sendline(self.mig_bg_kill_command)
+            session.close()
 
-        memorys = self.get_all_memorys(vm)
-        return set(filter(is_offline_memory, memorys))
+    def ping_pong_migrate(self, mig_type, sync, start_work=None,
+                          check_work=None):
 
-    def get_all_memorys(self, vm):
-        """Get all memorys detected in guest OS"""
-        mem_sys_path = "/sys/devices/system/memory"
-        cmd = "ls %s | grep memory" % mem_sys_path
-        session = self.get_session(vm)
-        output = session.cmd_output_safe(cmd, timeout=90)
-        return set([_ for _ in output.splitlines() if _])
+        """
+        ping pong migration test
+
+        :param mig_type: class MultihostMigration
+        :param sync: class SyncData
+        :param start_work: run sub test on src before migration
+        :param check_work: run sub test on dst after migration
+        """
+
+        while True:
+            if self.stop_migrate:
+                break
+            logging.info("ping pong migration...")
+            mig_type(self.test, self.params, self.env).migrate_wait(
+                [self.vm], self.srchost, self.dsthost,
+                start_work=start_work, check_work=check_work)
+            sync.sync(True, timeout=self.login_timeout)
+            vm = self.env.get_vm(self.params["main_vm"])
+            if vm.is_dead():
+                self.stop_migrate = True
+            elif self.migrate_count-1 == 0:
+                self.stop_migrate = True
+            else:
+                self.dsthost, self.srchost = self.srchost, self.dsthost
+                self.is_src = not self.is_src
+                start_work = None
+
+    @error.context_aware
+    def get_migration_info(self, vm):
+
+        """
+        get info after migration, focus on if keys in returned disc.
+
+        :param vm: vm object
+        """
+
+        error.context("Get 'xbzrle-cache/status/setup-time/downtime/"
+                      "total-time/ram' info after migration.",
+                      logging.info)
+        xbzrle_cache = vm.monitor.info("migrate").get("xbzrle-cache")
+        status = vm.monitor.info("migrate").get("status")
+        setup_time = vm.monitor.info("migrate").get("setup-time")
+        downtime = vm.monitor.info("migrate").get("downtime")
+        total_time = vm.monitor.info("migrate").get("total-time")
+        ram = vm.monitor.info("migrate").get("ram")
+        logging.info("Migration info:\nxbzrle-cache: %s\nstatus: %s\n"
+                     "setup-time: %s\ndowntime: %s\ntotal-time: "
+                     "%s\nram: %s" % (xbzrle_cache, status, setup_time,
+                                      downtime, total_time, ram))
+
+    @error.context_aware
+    def get_migration_capability(self, index=0):
+
+        """
+        Get the state of migrate-capability.
+
+        :param index: the index of capabilitys list.
+        """
+
+        if self.is_src:
+            for i in range(index, len(self.capabilitys)):
+                error.context("Get capability '%s' state."
+                              % self.capabilitys[i], logging.info)
+                vm = self.env.get_vm(self.params["main_vm"])
+                self.state = vm.monitor.get_migrate_capability(
+                    self.capabilitys[i])
+                if self.state != self.capabilitys_state[i]:
+                    raise error.TestFail(
+                        "The expected '%s' state: '%s',"
+                        " Actual result: '%s'." % (
+                            self.capabilitys[i],
+                            self.capabilitys_state[i],
+                            self.state))
+
+    @error.context_aware
+    def set_migration_capability(self, state, capability):
+
+        """
+        Set the capability of migrate to state.
+
+        :param state: Bool value of capability.
+        :param capability: capability which need to set.
+        """
+
+        if self.is_src:
+            error.context("Set '%s' state to '%s'." % (capability, state),
+                          logging.info)
+            vm = self.env.get_vm(self.params["main_vm"])
+            vm.monitor.set_migrate_capability(state, capability)
+
+    @error.context_aware
+    def get_migration_cache_size(self, index=0):
+
+        """
+        Get the xbzrle cache size.
+
+        :param index: the index of cache_size list
+        """
+
+        if self.is_src:
+            error.context("Try to get cache size.", logging.info)
+            vm = self.env.get_vm(self.params["main_vm"])
+            cache_size = vm.monitor.get_migrate_cache_size()
+            error.context("Get cache size: %s" % cache_size, logging.info)
+            if cache_size != int(self.cache_size[index]):
+                raise error.TestFail(
+                    "The expected cache size: %s,"
+                    " Actual result: %s." % (self.cache_size[index],
+                                             cache_size))
+
+    @error.context_aware
+    def set_migration_cache_size(self, value):
+
+        """
+        Set the cache size of migrate to value.
+
+        :param value: the cache size to set.
+        """
+
+        if self.is_src:
+            error.context("Set cache size to %s." % value, logging.info)
+            vm = self.env.get_vm(self.params["main_vm"])
+            vm.monitor.set_migrate_cache_size(value)
+
+    @error.context_aware
+    def get_migration_parameter(self, index=0):
+
+        """
+        Get the value of parameter.
+
+        :param index: the index of parameters list.
+        """
+
+        if self.is_src:
+            for i in range(index, len(self.parameters)):
+                error.context("Get parameter '%s' value."
+                              % self.parameters[i], logging.info)
+                vm = self.env.get_vm(self.params["main_vm"])
+                self.value = vm.monitor.get_migrate_parameter(
+                    self.parameters[i])
+                if int(self.value) != int(self.parameters_value[i]):
+                    raise error.TestFail(
+                        "The expected '%s' value: '%s',"
+                        " Actual result: '%s'." % (
+                            self.parameters[i],
+                            self.parameters_value[i],
+                            self.value))
+
+    @error.context_aware
+    def set_migration_parameter(self, index=0):
+
+        """
+        Set the value of parameter.
+
+        :param index: the index of parameters/parameters_value list.
+        """
+
+        if self.is_src:
+            for i in range(index, len(self.parameters)):
+                error.context("Set '%s' value to '%s'." % (
+                    self.parameters[i],
+                    self.parameters_value[i]), logging.info)
+                vm = self.env.get_vm(self.params["main_vm"])
+                vm.monitor.set_migrate_parameter(self.parameters[i],
+                                                 int(self.parameters_value[i]))
+
+    @error.context_aware
+    def set_migration_speed(self, value):
+
+        """
+        Set maximum speed (in bytes/sec) for migrations.
+
+        :param value: Speed in bytes/sec
+        """
+
+        if self.is_src:
+            error.context("Set migration speed to %s." % value, logging.info)
+            vm = self.env.get_vm(self.params["main_vm"])
+            vm.monitor.migrate_set_speed("%sB" % value)
+
+    @error.context_aware
+    def set_migration_downtime(self, value):
+
+        """
+        Set maximum tolerated downtime (in seconds) for migration.
+
+        :param value: maximum downtime (in seconds)
+        """
+
+        if self.is_src:
+            error.context("Set downtime to %s." % value, logging.info)
+            vm = self.env.get_vm(self.params["main_vm"])
+            vm.monitor.migrate_set_downtime(value)
+
+    @error.context_aware
+    def set_migration_cancel(self):
+
+        """
+        Cancel migration after it is beginning
+        """
+
+        if self.is_src:
+            error.context("Cancel migration.", logging.info)
+            vm = self.env.get_vm(self.params["main_vm"])
+            vm.monitor.cmd("migrate_cancel")
+
+    @error.context_aware
+    def get_migration_cancelled(self):
+
+        """
+        check the migration cancelled
+        """
+
+        if self.is_src:
+            vm = self.env.get_vm(self.params["main_vm"])
+            o = vm.monitor.info("migrate")
+            if isinstance(o, str):
+                return ("Migration status: cancelled" in o or
+                        "Migration status: canceled" in o)
+            else:
+                return (o.get("status") == "cancelled" or
+                        o.get("status") == "canceled")
+
+    @error.context_aware
+    def clean_up(self, kill_bg_cmd, vm):
+
+        """
+        kill background cmd on dst after migration
+
+        :param kill_bg_cmd: cmd for kill background test
+        :param vm:  vm object
+        """
+
+        error.context("Kill the background test by '%s' in guest"
+                      "." % kill_bg_cmd, logging.info)
+        session = vm.wait_for_login(timeout=self.login_timeout)
+        if session.cmd_status(self.check_running_cmd) != 0:
+            logging.info("The background test in guest is finished, "
+                         "no need to kill.")
+        else:
+            try:
+                s, o = session.cmd_status_output(kill_bg_cmd)
+                logging.info("The output after run kill command: %r" % o)
+                if "No such process" in o or "not found" in o \
+                        or "no running instance" in o:
+                    if session.cmd_status(self.check_running_cmd) != 0:
+                        logging.info("The background test in guest is "
+                                     "finished before kill it.")
+                elif s:
+                    raise error.TestFail("Failed to kill the background"
+                                         " test in guest.")
+            except (aexpect.ShellStatusError, aexpect.ShellTimeoutError):
+                pass
+        session.close()
+
+    @error.context_aware
+    def start_stress(self):
+
+        """
+        start stress test on src before migration
+        """
+
+        logging.info("Try to login guest before migration test.")
+        vm = self.env.get_vm(self.params["main_vm"])
+        session = vm.wait_for_login(timeout=self.login_timeout)
+        error.context("Do stress test before migration.", logging.info)
+        bg = utils.InterruptedThread(
+            utils_test.run_virt_sub_test,
+            args=(self.test, self.params, self.env,),
+            kwargs={"sub_type": self.bg_stress_test})
+        bg.start()
+        time.sleep(self.WAIT_SHORT)
+
+        def check_running():
+            return session.cmd_status(self.check_running_cmd) == 0
+
+        if self.check_running_cmd:
+            if not utils_misc.wait_for(check_running, timeout=360):
+                raise error.TestFail("Failed to start %s in guest." %
+                                     self.bg_stress_test)
+
+    @error.context_aware
+    def install_stressapptest(self):
+
+        """
+        install stressapptest
+        """
+
+        vm = self.env.get_vm(self.params["main_vm"])
+        session = vm.wait_for_login(timeout=self.login_timeout)
+        app_repo = "git clone https://github.com/stressapptest/" \
+                   "stressapptest.git"
+        stressapptest_insatll_cmd = "rm -rf stressapptest " \
+                                    "&& %s" \
+                                    " && cd stressapptest " \
+                                    "&& ./configure " \
+                                    "&& make " \
+                                    "&& make install" % app_repo
+        stressapptest_insatll_cmd = \
+            self.params.get("stressapptest_insatll_cmd",
+                            stressapptest_insatll_cmd)
+        error.context("Install stressapptest.", logging.info)
+        s, o = session.cmd_status_output(stressapptest_insatll_cmd)
+        session.close()
+        if s:
+            raise error.TestError("Failed to install stressapptest "
+                                  "in guest: '%s'" % o)
